+++ agent/config/config_test.go
@@ -14,6 +14,7 @@
 package config
 
 import (
+	"encoding/json"
 	"errors"
 	"os"
 	"reflect"
@@ -80,37 +81,56 @@ func TestBrokenEC2MetadataEndpoint(t *testing.T) {
 }
 
 func TestEnvironmentConfig(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_CLUSTER", "myCluster")
+	defer os.Unsetenv("ECS_CLUSTER")
 	os.Setenv("ECS_RESERVED_PORTS_UDP", "[42,99]")
+	defer os.Unsetenv("ECS_RESERVED_PORTS_UDP")
 	os.Setenv("ECS_RESERVED_MEMORY", "20")
+	defer os.Unsetenv("ECS_RESERVED_MEMORY")
 	os.Setenv("ECS_CONTAINER_STOP_TIMEOUT", "60s")
+	defer os.Unsetenv("ECS_CONTAINER_STOP_TIMEOUT")
 	os.Setenv("ECS_AVAILABLE_LOGGING_DRIVERS", "[\""+string(dockerclient.SyslogDriver)+"\"]")
+	defer os.Unsetenv("ECS_AVAILABLE_LOGGING_DRIVERS")
 	os.Setenv("ECS_SELINUX_CAPABLE", "true")
+	defer os.Unsetenv("ECS_SELINUX_CAPABLE")
 	os.Setenv("ECS_APPARMOR_CAPABLE", "true")
+	defer os.Unsetenv("ECS_APPARMOR_CAPABLE")
 	os.Setenv("ECS_DISABLE_PRIVILEGED", "true")
+	defer os.Unsetenv("ECS_DISABLE_PRIVILEGED")
 	os.Setenv("ECS_ENGINE_TASK_CLEANUP_WAIT_DURATION", "90s")
+	defer os.Unsetenv("ECS_ENGINE_TASK_CLEANUP_WAIT_DURATION")
 	os.Setenv("ECS_ENABLE_TASK_IAM_ROLE", "true")
+	defer os.Unsetenv("ECS_ENABLE_TASK_IAM_ROLE")
 	os.Setenv("ECS_ENABLE_TASK_IAM_ROLE_NETWORK_HOST", "true")
+	defer os.Unsetenv("ECS_ENABLE_TASK_IAM_ROLE_NETWORK_HOST")
 	os.Setenv("ECS_DISABLE_IMAGE_CLEANUP", "true")
+	defer os.Unsetenv("ECS_DISABLE_IMAGE_CLEANUP")
 	os.Setenv("ECS_IMAGE_CLEANUP_INTERVAL", "2h")
+	defer os.Unsetenv("ECS_IMAGE_CLEANUP_INTERVAL")
 	os.Setenv("ECS_IMAGE_MINIMUM_CLEANUP_AGE", "30m")
+	defer os.Unsetenv("ECS_IMAGE_MINIMUM_CLEANUP_AGE")
 	os.Setenv("ECS_NUM_IMAGES_DELETE_PER_CYCLE", "2")
+	defer os.Unsetenv("ECS_NUM_IMAGES_DELETE_PER_CYCLE")
 	os.Setenv("ECS_INSTANCE_ATTRIBUTES", "{\"my_attribute\": \"testing\"}")
+	defer os.Unsetenv("ECS_INSTANCE_ATTRIBUTES")
 	os.Setenv("ECS_ENABLE_TASK_ENI", "true")
+	defer os.Unsetenv("ECS_ENABLE_TASK_ENI")
+	additionalLocalRoutesJSON := `["1.2.3.4/22","5.6.7.8/32"]`
+	os.Setenv("ECS_AWSVPC_ADDITIONAL_LOCAL_ROUTES", additionalLocalRoutesJSON)
+	defer os.Unsetenv("ECS_AWSVPC_ADDITIONAL_LOCAL_ROUTES")
 
 	conf, err := environmentConfig()
-	assert.Nil(t, err)
+	assert.NoError(t, err)
 	assert.Equal(t, "myCluster", conf.Cluster)
 	assert.Equal(t, 2, len(conf.ReservedPortsUDP))
 	assert.Contains(t, conf.ReservedPortsUDP, uint16(42))
 	assert.Contains(t, conf.ReservedPortsUDP, uint16(99))
 	assert.Equal(t, uint16(20), conf.ReservedMemory)
-
 	expectedDuration, _ := time.ParseDuration("60s")
 	assert.Equal(t, expectedDuration, conf.DockerStopTimeout)
-
 	assert.Equal(t, []dockerclient.LoggingDriver{dockerclient.SyslogDriver}, conf.AvailableLoggingDrivers)
-
 	assert.True(t, conf.PrivilegedDisabled)
 	assert.True(t, conf.SELinuxCapable, "Wrong value for SELinuxCapable")
 	assert.True(t, conf.AppArmorCapable, "Wrong value for AppArmorCapable")
@@ -118,15 +138,19 @@ func TestEnvironmentConfig(t *testing.T) {
 	assert.True(t, conf.TaskIAMRoleEnabledForNetworkHost, "Wrong value for TaskIAMRoleEnabledForNetworkHost")
 	assert.True(t, conf.ImageCleanupDisabled, "Wrong value for ImageCleanupDisabled")
 	assert.True(t, conf.TaskENIEnabled, "Wrong value for TaskNetwork")
-
-	assert.Equal(t, (30 * time.Minute), conf.MinimumImageDeletionAge)
-	assert.Equal(t, (2 * time.Hour), conf.ImageCleanupInterval)
+	assert.Equal(t, 30*time.Minute, conf.MinimumImageDeletionAge)
+	assert.Equal(t, 2*time.Hour, conf.ImageCleanupInterval)
 	assert.Equal(t, 2, conf.NumImagesToDeletePerCycle)
 	assert.Equal(t, "testing", conf.InstanceAttributes["my_attribute"])
-	assert.Equal(t, (90 * time.Second), conf.TaskCleanupWaitDuration)
+	assert.Equal(t, 90*time.Second, conf.TaskCleanupWaitDuration)
+	serializedAdditionalLocalRoutesJSON, err := json.Marshal(conf.AWSVPCAdditionalLocalRoutes)
+	assert.NoError(t, err, "should marshal additional local routes")
+	assert.Equal(t, additionalLocalRoutesJSON, string(serializedAdditionalLocalRoutesJSON))
 }
 
 func TestTrimWhitespace(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_CLUSTER", "default \r")
 	os.Setenv("ECS_ENGINE_AUTH_TYPE", "dockercfg\r")
 
@@ -153,6 +177,8 @@ func TestTrimWhitespace(t *testing.T) {
 }
 
 func TestConfigBoolean(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_DISABLE_METRICS", "true")
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
 	if err != nil {
@@ -253,6 +279,8 @@ func TestValidFormatParseEnvVariableDuration(t *testing.T) {
 }
 
 func TestInvalidTaskCleanupTimeout(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_ENGINE_TASK_CLEANUP_WAIT_DURATION", "1s")
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
 	if err != nil {
@@ -267,6 +295,8 @@ func TestInvalidTaskCleanupTimeout(t *testing.T) {
 }
 
 func TestTaskCleanupTimeout(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_ENGINE_TASK_CLEANUP_WAIT_DURATION", "10m")
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
 	if err != nil {
@@ -281,6 +311,8 @@ func TestTaskCleanupTimeout(t *testing.T) {
 }
 
 func TestInvalidReservedMemory(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_RESERVED_MEMORY", "-1")
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
 	if err != nil {
@@ -295,6 +327,8 @@ func TestInvalidReservedMemory(t *testing.T) {
 }
 
 func TestReservedMemory(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_RESERVED_MEMORY", "1")
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
 	if err != nil {
@@ -309,6 +343,8 @@ func TestReservedMemory(t *testing.T) {
 }
 
 func TestTaskIAMRoleEnabled(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_ENABLE_TASK_IAM_ROLE", "true")
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
 	if err != nil {
@@ -321,6 +357,8 @@ func TestTaskIAMRoleEnabled(t *testing.T) {
 }
 
 func TestTaskIAMRoleForHostNetworkEnabled(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_ENABLE_TASK_IAM_ROLE_NETWORK_HOST", "true")
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
 	if err != nil {
@@ -333,6 +371,8 @@ func TestTaskIAMRoleForHostNetworkEnabled(t *testing.T) {
 }
 
 func TestCredentialsAuditLogFile(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	dummyLocation := "/foo/bar.log"
 	os.Setenv("ECS_AUDIT_LOGFILE", dummyLocation)
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
@@ -346,6 +386,8 @@ func TestCredentialsAuditLogFile(t *testing.T) {
 }
 
 func TestCredentialsAuditLogDisabled(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_AUDIT_LOGFILE_DISABLED", "true")
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
 	if err != nil {
@@ -358,6 +400,8 @@ func TestCredentialsAuditLogDisabled(t *testing.T) {
 }
 
 func TestImageCleanupMinimumInterval(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_IMAGE_CLEANUP_INTERVAL", "1m")
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
 	if err != nil {
@@ -370,6 +414,8 @@ func TestImageCleanupMinimumInterval(t *testing.T) {
 }
 
 func TestImageCleanupMinimumNumImagesToDeletePerCycle(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_NUM_IMAGES_DELETE_PER_CYCLE", "-1")
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
 	if err != nil {
@@ -382,9 +428,18 @@ func TestImageCleanupMinimumNumImagesToDeletePerCycle(t *testing.T) {
 }
 
 func TestAWSVPCBlockInstanceMetadata(t *testing.T) {
+	os.Setenv("AWS_DEFAULT_REGION", "foo-bar-1")
+	defer os.Unsetenv("AWS_DEFAULT_REGION")
 	os.Setenv("ECS_AWSVPC_BLOCK_IMDS", "true")
 	defer os.Unsetenv("ECS_AWSVPC_BLOCK_IMDS")
 	cfg, err := NewConfig(ec2.NewBlackholeEC2MetadataClient())
 	assert.NoError(t, err)
 	assert.True(t, cfg.AWSVPCBlockInstanceMetdata)
+}
+
+func TestInvalidAWSVPCAdditionalLocalRoutes(t *testing.T) {
+	os.Setenv("ECS_AWSVPC_ADDITIONAL_LOCAL_ROUTES", `["300.300.300.300/64"]`)
+	defer os.Unsetenv("ECS_AWSVPC_ADDITIONAL_LOCAL_ROUTES")
+	_, err := environmentConfig()
+	assert.Error(t, err)
 }