+++ agent/config/types.go
@@ -18,6 +18,7 @@ import (
 	"time"
 
 	"github.com/aws/amazon-ecs-agent/agent/engine/dockerclient"
+	cnitypes "github.com/containernetworking/cni/pkg/types"
 )
 
 type Config struct {
@@ -167,6 +168,18 @@ type Config struct {
 	// AWSVPCBlockInstanceMetdata specifies if InstanceMetadata endpoint should be blocked
 	// for tasks that are launched with network mode "awsvpc" when ECS_AWSVPC_BLOCK_IMDS=true
 	AWSVPCBlockInstanceMetdata bool
+
+	// OverrideAWSVPCLocalIPv4Address overrides the local IPv4 address chosen
+	// for a task using the `awsvpc` networking mode. Using this configuration
+	// will limit you to running one `awsvpc` task at a time. IPv4 addresses
+	// must be specified in decimal-octet form and also specify the subnet
+	// size (e.g., "169.254.172.42/22").
+	OverrideAWSVPCLocalIPv4Address *cnitypes.IPNet
+
+	// AWSVPCAdditionalLocalRoutes allows the specification of routing table
+	// entries that will be added in the task's network namespace via the
+	// instance bridge interface rather than via the ENI.
+	AWSVPCAdditionalLocalRoutes []cnitypes.IPNet
 }
 
 // SensitiveRawMessage is a struct to store some data that should not be logged